Move sensor type creation code to st2common
