+++ st2reactor/st2reactor/bootstrap/sensorsregistrar.py
@@ -19,8 +19,9 @@ import six
 
 from st2common import log as logging
 from st2common.bootstrap.base import ResourceRegistrar
-import st2reactor.container.utils as container_utils
 import st2common.content.utils as content_utils
+from st2common.models.api.reactor import SensorTypeAPI
+from st2common.persistence.reactor import SensorType
 
 __all__ = [
     'SensorsRegistrar',
@@ -112,41 +113,33 @@ class SensorsRegistrar(ResourceRegistrar):
         LOG.debug('Loading sensor from %s.', sensor_metadata_file_path)
         metadata = self._meta_loader.load(file_path=sensor_metadata_file_path)
 
-        class_name = metadata.get('class_name', None)
         entry_point = metadata.get('entry_point', None)
-        description = metadata.get('description', None)
-        trigger_types = metadata.get('trigger_types', [])
-        poll_interval = metadata.get('poll_interval', None)
-        enabled = metadata.get('enabled', True)
 
-        # Add pack to each trigger type item
-        for trigger_type in trigger_types:
-            trigger_type['pack'] = pack
+        if not entry_point:
+            raise ValueError('Sensor definition missing entry_point')
 
-        # Add TrigerType models to the DB
-        trigger_type_dbs = container_utils.add_trigger_models(trigger_types=trigger_types)
+        sensors_dir = os.path.dirname(sensor_metadata_file_path)
+        sensor_file_path = os.path.join(sensors_dir, entry_point)
+        artifact_uri = 'file://%s' % (sensor_file_path)
+        metadata['artifact_uri'] = artifact_uri
+        metadata['entry_point'] = entry_point
 
-        # Populate a list of references belonging to this sensor
-        trigger_type_refs = []
-        for trigger_type_db, _ in trigger_type_dbs:
-            ref_obj = trigger_type_db.get_reference()
-            trigger_type_ref = ref_obj.ref
-            trigger_type_refs.append(trigger_type_ref)
+        sensor_api = SensorTypeAPI(**metadata)
+        sensor_model = sensor_api.to_model()
 
-        if entry_point and class_name:
-            sensors_dir = os.path.dirname(sensor_metadata_file_path)
-            sensor_file_path = os.path.join(sensors_dir, entry_point)
-            # Add Sensor model to the DB
-            sensor_obj = {
-                'name': class_name,
-                'description': description,
-                'class_name': class_name,
-                'file_path': sensor_file_path,
-                'trigger_types': trigger_type_refs,
-                'poll_interval': poll_interval,
-                'enabled': enabled
-            }
-            container_utils.add_sensor_model(pack=pack, sensor=sensor_obj)
+        sensor_types = SensorType.query(pack=sensor_model.pack, name=sensor_model.name)
+        if len(sensor_types) >= 1:
+            sensor_type = sensor_types[0]
+            LOG.debug('Found existing sensor id:%s with name:%s. Will update it.',
+                      sensor_type.id, sensor_type.name)
+            sensor_model.id = sensor_type.id
+
+        try:
+            sensor_model = SensorType.add_or_update(sensor_model)
+        except:
+            LOG.exception('Failed creating sensor model for %s', sensor)
+
+        return sensor_model
 
 
 def register_sensors(packs_base_paths=None, pack_dir=None):
